feat: extract object execution from jdbcTemplate
