+++ jdbc/src/main/java/nextstep/jdbc/AbstractExecution.java
@@ -2,11 +2,16 @@ package nextstep.jdbc;
 
 public abstract class AbstractExecution<T> implements Execution<T> {
 
-    protected final String sql;
-    private final Object[] arguments;
+    private final String sql;
+    protected final Object[] arguments;
 
     public AbstractExecution(String sql, Object[] arguments) {
         this.sql = sql;
         this.arguments = arguments;
+    }
+
+    @Override
+    public String getSql() {
+        return sql;
     }
 }
