+++ jdbc/src/main/java/nextstep/jdbc/JdbcConnector.java
@@ -4,8 +4,12 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import javax.sql.DataSource;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class JdbcConnector {
+
+    private static final Logger log = LoggerFactory.getLogger(JdbcConnector.class);
 
     private DataSource dataSource;
 
@@ -18,7 +22,7 @@ public class JdbcConnector {
              PreparedStatement statement = connection.prepareStatement(execution.getSql())) {
             return execution.execute(statement);
         } catch (SQLException e) {
-//            log.info(e.getMessage());
+            log.info(e.getMessage());
             throw new RuntimeException(e);
         }
     }
