+++ jdbc/src/main/java/nextstep/jdbc/ObjectExecution.java
@@ -1,0 +1,27 @@
+package nextstep.jdbc;
+
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+public class ObjectExecution<T> extends AbstractExecution<T> {
+
+    private final RowMapper<T> rowMapper;
+
+    public ObjectExecution(String sql, Object[] arguments, RowMapper<T> rowMapper) {
+        super(sql, arguments);
+        this.rowMapper = rowMapper;
+    }
+
+    @Override
+    public T execute(PreparedStatement statement) throws SQLException {
+        for (int i = 0; i < arguments.length; i++) {
+            statement.setObject(i + 1, arguments[i]);
+        }
+        ResultSet resultSet = statement.executeQuery();
+        if (resultSet.next()) {
+            return rowMapper.rowMap(resultSet, resultSet.getRow());
+        }
+        return null;
+    }
+}
